requirements.txt
pandas>=2.2
requests>=2.31
requests-cache>=1.2
tqdm>=4.66
tenacity>=8.2
pyarrow>=15.0
SQLAlchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0
